pyproject.toml
[tool.poetry]
name = "summarization-api"
version = "0.1.0"
description = "AI Summarization API with AWS Lambda and API Gateway"
authors = ["Abraham Linsenmayer"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "^2.100.0"
constructs = "^10.0.0"
boto3 = "^1.29.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
